Add ability to get all repos by org or user
Now in addition to being able to a specific repo user the format
user/repo, you can just pass org or user name to get consolidated stats
for all of their repos that are not forks.

This does bring about another issue where if there are a lot of repos,
we can start to run up against GitHub's unauthenticated rate limiting. I
was hoping to keep things simple enough to not have to add auth, so for
now at least, this will just cause a failure to collect the stats and
the graph will be flat.